fix: fixing bugs
fixing some bugs and componetization
