+++ src/components/Tabber/Tabber.module.scss
@@ -1,7 +1,8 @@
 .social {
-  display: flex;
-
-  justify-content: start;
-
-  padding: 1.5rem;
-}
+    display: flex;
+  
+    justify-content: flex-start;
+  
+    padding: 1.6rem;
+  }
+
