+++ src/components/Tabber/Tabber.stories.tsx
@@ -1,0 +1,9 @@
+import { Story } from '@ladle/react';
+
+import { networkListMock } from './networkListMock';
+
+import Tabber from './Tabber';
+
+export const TabberComponent: Story = () => {
+  return <Tabber socialList={networkListMock}></Tabber>;
+};
