+++ src/components/Tabber/components/Tabs/TabsTabber.module.scss
@@ -3,9 +3,9 @@
   flex-direction: row;
 
   align-items: center;
-  justify-content: center; /* Adicionado para alinhamento horizontal */
+  justify-content: center; 
 
-  color: var(--m-3-sys-light-primary, #868687);
+  color:  #868687;
   font-size: 0.875rem;
   line-height: 1.25rem;
 
@@ -13,19 +13,19 @@
   letter-spacing: 0.00625rem;
 
   &:hover {
-    background-color: #6adafb;
+    background-color: #dce2e4;
   }
 
   &.selected {
-    color: var(--m-3-sys-light-primary, #7033e0);
+    color:  #7033e0;
 
-    border-bottom: 0.2rem solid currentColor;
+    border-bottom: 2px solid currentColor;
   }
 
   .icon {
-    width: 1.5rem;
-    height: 1.5rem;
+    width: 1.6rem;
+    height: 1.6rem;
 
-    padding: 0.75rem 0.5rem;
+    padding: 0.8rem 0.6rem;
   }
 }
